api: do not instanciate the Api with the curren Flask app
It's more explicit to have (1) the 'app' and (2) the 'api' instances in two
modules, resp. 'webapp.py' and 'webapi.py'. With flask-restplus, the Flask app
could be configured when the Api was instanciated. Thus, Flas app
configurations were set after that were not taken into account.

You'll initialize it with the lazy
way, i.e.

    api.init_app(app)

in a dedicated Python script, such as 'wsgi.py'.