ui: fix the wrong cluster id selection in centroids time series plot
Until now, the i-th cluster was considered as the i-th key in the JSON
structure got from the API. However dict keys are not ordered, so the cluster
ids may be unrelated to the data they refer to. This commit fix this issue, by
implementing a `clusters` dictionary in which keys refers to good ids.