tasks: initialize a city.py module to factorize bordeaux and lyon pipeline
This commit introduces a new module that factorize lyon.py and
bordeaux.py, so as to accomplish the pipeline first steps (station zip
file downloading and unzipping). For the moment, the download URLs are
hard-written as module parameters; a further module improvement may be
to design a params_factory function as in lyon.py.